Add some copts from @airbus-seclab (#111)

+++ voidstar/algos_2d/algo_2d.h
@@ -9,20 +9,20 @@ class Algo2D : public Algorithm {
   static constexpr char kSceneType[] = "Scene2D";
   virtual std::string sceneType() const final { return kSceneType; }
 
-  virtual bool apply(Floats& vertices, Floats& colors, size_t width,
-                     size_t height) = 0;
+  virtual bool apply(Floats& vertices, Floats& colors, u32 width,
+                     u32 height) = 0;
 
  protected:
-  size_t make_vertices(Floats& vertices, size_t width, size_t height) {
+  u32 make_vertices(Floats& vertices, u32 width, u32 height) {
     const float w = static_cast<float>(width) / 2;
     const float h = static_cast<float>(height) / 2;
 
     size_t pos = 0;
-    for (size_t y = 0; y < height; ++y)
-      for (size_t x = 0; x < width; ++x) {
+    for (u32 y = 0; y < height; ++y)
+      for (u32 x = 0; x < width; ++x) {
         vertices[pos++] = (static_cast<float>(x) - w) / w;
         vertices[pos++] = (static_cast<float>(y) - h) / h;
       }
-    return vertices.size() / 2;
+    return static_cast<u32>(vertices.size()) / 2;
   }
 };
