+++ voidstar/arguments.h
@@ -4,6 +4,7 @@
 #include <memory>
 
 #include "voidstar/registrar.h"
+#include "voidstar/types.h"
 
 struct Arguments {
   Arguments() {}
@@ -12,15 +13,15 @@ struct Arguments {
   std::string name = "void*";
   std::string algo = "Algo3DCubeContiFrebet";
   std::string manager = "GLFW3Manager";
-  size_t width = 800;
-  size_t height = 600;
-  size_t range_begin = 0;
-  size_t range_end = 0;
+  u32 width = 800;
+  u32 height = 600;
+  u32 range_begin = 0;
+  u32 range_end = 0;
   bool fullscreen = false;
   bool keep_chrome = false;
-  size_t sliding_window_length = 37 * 1024;
-  size_t sliding_step = 1024;
-  size_t sliding_step_factor = 1;
+  u32 sliding_window_length = 37 * 1024;
+  u32 sliding_step = 1024;
+  u32 sliding_step_factor = 1;
   bool move_window = false;
   bool spin_shape = true;
   bool exit_at_fin = false;
